+++ packages/code/src/Syntax/index.ts
@@ -1,0 +1,1 @@
+export { default as Syntax } from './Syntax';