+++ packages/icon-button/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @leafygreen-ui/icon-button
 
+## 15.0.6
+
+### Patch Changes
+
+- bf2fedf6d: Version bumps lib
+- Updated dependencies [51c544e2e]
+- Updated dependencies [bf2fedf6d]
+  - @leafygreen-ui/a11y@1.4.0
+  - @leafygreen-ui/leafygreen-provider@3.1.1
+
+## 15.0.5
+
+### Patch Changes
+
+- c1c259036: Exporting types
+
 ## 15.0.4
 
 ### Patch Changes